requirements.txt
boto3
requests
python-dotenv
aws-wsgi
flask